refactor: organize output processing into src/reports module
- Created src/reports module to centralize all output-related functionality
- Moved SARIF report generation from src/sarif.rs to src/reports/sarif.rs
- Extracted filename generation functions from main.rs to src/reports/filename.rs
- Extracted markdown generation from response.rs to src/reports/markdown.rs
- Extracted analysis summary from response.rs to src/reports/summary.rs
- Moved output directory validation from args.rs to src/reports/validation.rs
- Updated all imports and references to use the new module structure
- All tests passing with improved code organization
