+++ src/args.rs
@@ -1,7 +1,5 @@
 use anyhow::Result;
 use clap::Parser;
-use std::fs;
-use std::io::Write;
 use std::path::PathBuf;
 
 #[derive(Parser, Debug)]
@@ -56,36 +54,9 @@ pub struct Args {
     pub language: String,
 }
 
-pub fn validate_output_directory(output_dir: &PathBuf) -> Result<()> {
-    if !output_dir.exists() {
-        fs::create_dir_all(output_dir)
-            .map_err(|e| anyhow::anyhow!("ディレクトリの作成に失敗: {}", e))?;
-    }
-
-    let mut test_file_path = output_dir.clone();
-    test_file_path.push(".parsentry_write_test");
-
-    match fs::File::create(&test_file_path) {
-        Ok(mut file) => {
-            if let Err(e) = file.write_all(b"test") {
-                let _ = fs::remove_file(&test_file_path);
-                return Err(anyhow::anyhow!("書き込み権限がありません: {}", e));
-            }
-            drop(file);
-            fs::remove_file(&test_file_path)
-                .map_err(|e| anyhow::anyhow!("テストファイルの削除に失敗: {}", e))?;
-        }
-        Err(e) => {
-            return Err(anyhow::anyhow!("ファイル作成権限がありません: {}", e));
-        }
-    }
-
-    Ok(())
-}
-
 pub fn validate_args(args: &Args) -> Result<()> {
     if let Some(output_dir) = &args.output_dir {
-        if let Err(e) = validate_output_directory(output_dir) {
+        if let Err(e) = crate::reports::validate_output_directory(output_dir) {
             eprintln!(
                 "❌ 出力ディレクトリのチェックに失敗: {}: {}",
                 output_dir.display(),
