+++ src/lib.rs
@@ -1,13 +1,11 @@
 pub mod analyzer;
 pub mod args;
 pub mod file_classifier;
-pub mod language;
 pub mod locales;
 pub mod parser;
 pub mod pattern_generator;
 pub mod prompts;
 pub mod repo;
-pub mod repo_clone;
+pub mod reports;
 pub mod response;
-pub mod sarif;
 pub mod security_patterns;
