+++ src/prompts/evaluator.rs
@@ -1,4 +1,4 @@
-use crate::language::Language;
+use crate::locales::Language;
 use crate::locales;
 
 pub fn get_evaluator_prompt_template(lang: &Language) -> &'static str {
