+++ src/response.rs
@@ -1,7 +1,5 @@
 use serde::{Deserialize, Serialize};
 use serde_json::json;
-use std::collections::HashMap;
-use std::path::PathBuf;
 
 #[derive(Debug, Clone, Serialize, Deserialize, Hash, Eq, PartialEq)]
 pub enum VulnType {
@@ -418,166 +416,7 @@ impl Response {
     }
 
     pub fn to_markdown(&self) -> String {
-        let mut md = String::new();
-        
-        // Enhanced title with file and pattern information
-        let title = if let (Some(file_path), Some(pattern)) = (&self.file_path, &self.pattern_description) {
-            format!("# Security Analysis: {} - {}", 
-                file_path.split('/').last().unwrap_or(file_path), 
-                pattern)
-        } else if let Some(file_path) = &self.file_path {
-            format!("# Security Analysis: {}", 
-                file_path.split('/').last().unwrap_or(file_path))
-        } else {
-            "# Security Analysis Report".to_string()
-        };
-        md.push_str(&title);
-        md.push_str("\n\n");
-
-        // File information section
-        if let Some(file_path) = &self.file_path {
-            md.push_str("## ファイル情報\n\n");
-            md.push_str(&format!("- **ファイルパス**: `{}`\n", file_path));
-            if let Some(pattern) = &self.pattern_description {
-                md.push_str(&format!("- **検出パターン**: {}\n", pattern));
-            }
-            md.push_str("\n");
-        }
-
-        let confidence_badge = match self.confidence_score {
-            90..=100 => "![高信頼度](https://img.shields.io/badge/信頼度-高-red)",
-            70..=89 => "![中高信頼度](https://img.shields.io/badge/信頼度-中高-orange)",
-            50..=69 => "![中信頼度](https://img.shields.io/badge/信頼度-中-yellow)",
-            30..=49 => "![中低信頼度](https://img.shields.io/badge/信頼度-中低-green)",
-            _ => "![低信頼度](https://img.shields.io/badge/信頼度-低-blue)",
-        };
-        md.push_str(&format!(
-            "{} **信頼度スコア: {}**\n\n",
-            confidence_badge, self.confidence_score
-        ));
-
-        if !self.vulnerability_types.is_empty() {
-            md.push_str("## 脆弱性タイプ\n\n");
-            for vuln_type in &self.vulnerability_types {
-                md.push_str(&format!("- `{:?}`\n", vuln_type));
-            }
-            md.push('\n');
-        }
-
-        md.push_str("## PAR Policy Analysis\n\n");
-
-        if !self.par_analysis.principals.is_empty() {
-            md.push_str("### Principals (データ源)\n\n");
-            for principal in &self.par_analysis.principals {
-                md.push_str(&format!(
-                    "- **{}**: {:?}\n",
-                    principal.identifier, principal.trust_level
-                ));
-                md.push_str(&format!("  - Context: {}\n", principal.source_context));
-                md.push_str(&format!(
-                    "  - Risk Factors: {}\n",
-                    principal.risk_factors.join(", ")
-                ));
-            }
-            md.push('\n');
-        }
-
-        if !self.par_analysis.actions.is_empty() {
-            md.push_str("### Actions (セキュリティ制御)\n\n");
-            for action in &self.par_analysis.actions {
-                md.push_str(&format!(
-                    "- **{}**: {:?}\n",
-                    action.identifier, action.implementation_quality
-                ));
-                md.push_str(&format!("  - Function: {}\n", action.security_function));
-                md.push_str(&format!(
-                    "  - Weaknesses: {}\n",
-                    action.detected_weaknesses.join(", ")
-                ));
-                md.push_str(&format!(
-                    "  - Bypass Vectors: {}\n",
-                    action.bypass_vectors.join(", ")
-                ));
-            }
-            md.push('\n');
-        }
-
-        if !self.par_analysis.resources.is_empty() {
-            md.push_str("### Resources (操作対象)\n\n");
-            for resource in &self.par_analysis.resources {
-                md.push_str(&format!(
-                    "- **{}**: {:?}\n",
-                    resource.identifier, resource.sensitivity_level
-                ));
-                md.push_str(&format!("  - Operation: {}\n", resource.operation_type));
-                md.push_str(&format!(
-                    "  - Protection: {}\n",
-                    resource.protection_mechanisms.join(", ")
-                ));
-            }
-            md.push('\n');
-        }
-
-        if !self.par_analysis.policy_violations.is_empty() {
-            md.push_str("### Policy Violations\n\n");
-            for violation in &self.par_analysis.policy_violations {
-                md.push_str(&format!(
-                    "#### {}: {}\n\n",
-                    violation.rule_id, violation.rule_description
-                ));
-                md.push_str(&format!("- **Path**: {}\n", violation.violation_path));
-                md.push_str(&format!("- **Severity**: {}\n", violation.severity));
-                md.push_str(&format!(
-                    "- **Confidence**: {:.2}\n\n",
-                    violation.confidence
-                ));
-            }
-        }
-
-        // Source code sections
-        if let Some(matched_code) = &self.matched_source_code {
-            if !matched_code.trim().is_empty() {
-                md.push_str("## マッチしたソースコード\n\n");
-                md.push_str("```code\n");
-                md.push_str(matched_code);
-                md.push_str("\n```\n\n");
-            }
-        }
-
-        md.push_str("## 詳細解析\n\n");
-        md.push_str(&self.analysis);
-        md.push_str("\n\n");
-
-        if !self.poc.is_empty() {
-            md.push_str("## PoC（概念実証コード）\n\n");
-            md.push_str("```text\n");
-            md.push_str(&self.poc);
-            md.push_str("\n```\n\n");
-        }
-
-        if !self.remediation_guidance.policy_enforcement.is_empty() {
-            md.push_str("## 修復ガイダンス\n\n");
-            for remediation in &self.remediation_guidance.policy_enforcement {
-                md.push_str(&format!("### {}\n\n", remediation.component));
-                md.push_str(&format!(
-                    "- **Required**: {}\n",
-                    remediation.required_improvement
-                ));
-                md.push_str(&format!(
-                    "- **Guidance**: {}\n",
-                    remediation.specific_guidance
-                ));
-                md.push_str(&format!("- **Priority**: {}\n\n", remediation.priority));
-            }
-        }
-
-        if !self.scratchpad.is_empty() {
-            md.push_str("## 解析ノート\n\n");
-            md.push_str(&self.scratchpad);
-            md.push_str("\n\n");
-        }
-
-        md
+        crate::reports::markdown::to_markdown(self)
     }
 }
 
@@ -612,148 +451,3 @@ impl Response {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct FileAnalysisResult {
-    pub file_path: PathBuf,
-    pub response: Response,
-    pub output_filename: String, // The generated filename used for the actual markdown file
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct AnalysisSummary {
-    pub results: Vec<FileAnalysisResult>,
-}
-
-impl AnalysisSummary {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn add_result(&mut self, file_path: PathBuf, response: Response, output_filename: String) {
-        self.results.push(FileAnalysisResult {
-            file_path,
-            response,
-            output_filename,
-        });
-    }
-
-    pub fn sort_by_confidence(&mut self) {
-        self.results.sort_by(|a, b| {
-            b.response
-                .confidence_score
-                .cmp(&a.response.confidence_score)
-        });
-    }
-
-    pub fn filter_by_min_confidence(&self, min_score: i32) -> Self {
-        Self {
-            results: self
-                .results
-                .iter()
-                .filter(|r| r.response.confidence_score >= min_score)
-                .cloned()
-                .collect(),
-        }
-    }
-
-    pub fn filter_by_vuln_types(&self, vuln_types: &[VulnType]) -> Self {
-        Self {
-            results: self
-                .results
-                .iter()
-                .filter(|r| {
-                    r.response
-                        .vulnerability_types
-                        .iter()
-                        .any(|vt| vuln_types.contains(vt))
-                })
-                .cloned()
-                .collect(),
-        }
-    }
-
-    pub fn to_markdown(&self) -> String {
-        let mut md = String::new();
-        md.push_str("# PAR Security Analysis Summary Report\n\n");
-
-        md.push_str("## 概要\n\n");
-        md.push_str("| ファイル | 脆弱性タイプ | 信頼度 | Policy Violations |\n");
-        md.push_str("|---------|------------|--------|------------------|\n");
-
-        for result in &self.results {
-            if result.response.confidence_score > 0 {
-                let confidence_level = match result.response.confidence_score {
-                    90..=100 => "🔴 高",
-                    70..=89 => "🟠 中高",
-                    50..=69 => "🟡 中",
-                    30..=49 => "🟢 中低",
-                    _ => "🔵 低",
-                };
-
-                let vuln_types = result
-                    .response
-                    .vulnerability_types
-                    .iter()
-                    .map(|vt| format!("{:?}", vt))
-                    .collect::<Vec<_>>()
-                    .join(", ");
-
-                let violations = result
-                    .response
-                    .par_analysis
-                    .policy_violations
-                    .iter()
-                    .map(|v| v.rule_id.clone())
-                    .collect::<Vec<_>>()
-                    .join(", ");
-
-                // Create display name from filename + pattern if available
-                let display_name = if let Some(pattern) = &result.response.pattern_description {
-                    format!("{} ({})", 
-                        result.file_path.file_name().unwrap_or_default().to_string_lossy(),
-                        pattern)
-                } else {
-                    result.file_path.file_name().unwrap_or_default().to_string_lossy().to_string()
-                };
-
-                md.push_str(&format!(
-                    "| [{}]({}) | {} | {} | {} |\n",
-                    display_name,
-                    result.output_filename,
-                    vuln_types,
-                    confidence_level,
-                    violations
-                ));
-            }
-        }
-
-        md.push_str("\n## Policy Violation Analysis\n\n");
-
-        let mut violation_count: HashMap<String, i32> = HashMap::new();
-        for result in &self.results {
-            for violation in &result.response.par_analysis.policy_violations {
-                *violation_count
-                    .entry(violation.rule_id.clone())
-                    .or_insert(0) += 1;
-            }
-        }
-
-        md.push_str("| Rule ID | 件数 | 説明 |\n");
-        md.push_str("|---------|------|------|\n");
-
-        for (rule_id, count) in violation_count.iter() {
-            // Find the first occurrence to get the description
-            let description = self
-                .results
-                .iter()
-                .flat_map(|r| &r.response.par_analysis.policy_violations)
-                .find(|v| v.rule_id == *rule_id)
-                .map(|v| v.rule_description.clone())
-                .unwrap_or_default();
-
-            md.push_str(&format!("| {} | {} | {} |\n", rule_id, count, description));
-        }
-
-        md
-    }
-}
